requirements.txt
django==2.0.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
